Set up the basic cli structure with cobra
This commit goes as far as setting up Cobra in the project using their
boilerplate and folder structure. This will need cleaning up as each
task has information on the licence and command descriptions which need
to be filled in.

Cobra is a library for creating CLI applications in Go.
@link https://github.com/spf13/cobra#getting-started

It provides a handy command for adding new commands.
@link https://github.com/spf13/cobra/blob/master/cobra/README.md

`cobra add <command>` in the Go project will scaffold a new command.