Look at restructuring the app
I don't really know how to set this up. I was hoping that each part of
the logic could be its own file but I think I fundamentally don't
understand how that is supposed to work. There seems to be a lot of
reading to understand how Go wants it to look.

In the mean time I have the working main.go and then the rest split up
ready if I need it.